requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.21
pillow>=9.0
tqdm>=4.60
pyyaml>=6.0
requests>=2.25
